#4234 extracted to Carto::StaticMapsURLHelper.new.url_for_static_map
